issue #9 vue版本后台管理：用户管理功能没有设置角色功能

+++ guns-dao/src/main/java/cn/enilu/guns/dao/system/RoleRepository.java
@@ -17,7 +17,7 @@ public interface RoleRepository extends PagingAndSortingRepository<Role,Long> {
     List roleTreeList();
 
     @Query(nativeQuery = true,value="SELECT r.id AS id, pId AS pId, NAME AS NAME, ( CASE WHEN (pId = 0 OR pId IS NULL) THEN 'true' ELSE 'false' END ) \"open\", ( CASE WHEN (r1.ID = 0 OR r1.ID IS NULL) THEN 'false' ELSE 'true' END ) AS checked FROM t_sys_role r LEFT JOIN ( SELECT ID FROM t_sys_role WHERE ID IN (?1)) r1 ON r.ID = r1.ID ORDER BY pId, num ASC")
-    List roleTreeListByRoleId(Integer[] ids);
+    List roleTreeListByRoleId(Long[] ids);
 
     List findByName(String roleName);
 }
